hv: vm: Move launch_vms and create_vm to common scope
launch_vms and create_vm need to be moved together as both of
them reference file-static structures. All related structures
and helpers accessing structures are also moved.

Tracked-On: #8830
Acked-by: Wang Yu1 <yu1.wang@intel.com>
