+++ hypervisor/include/common/vm.h
@@ -146,7 +146,12 @@ bool is_paused_vm(__unused const struct acrn_vm *vm);
 bool is_poweroff_vm(__unused const struct acrn_vm *vm);
 
 void arch_trigger_level_intr(__unused struct acrn_vm *vm,
-			__unused uint32_t irq, __unused bool assert);
+		__unused uint32_t irq, __unused bool assert);
+
+int32_t arch_init_vm(struct acrn_vm *vm, struct acrn_vm_config *vm_config);
+
+int32_t create_vm(uint16_t vm_id, uint64_t pcpu_bitmap, struct acrn_vm_config *vm_config, struct acrn_vm **rtn_vm);
+void launch_vms(uint16_t pcpu_id);
 
 #endif /* !ASSEMBLER */
